refactor: reorganize proxy module[]
