+++ src/modules/proxy/proxy.module.ts
@@ -1,9 +1,22 @@
 import { Module } from "@nestjs/common";
-import { SocketIoModule } from "./socketio/socketio.module";
-import { HttpModule } from "./http/http.module";
-import { WebSocketModule } from "./websocket/websocket.module";
+import { HttpController } from "./controllers/http.controller";
+import { HttpService } from "./services/http.service";
+import { SocketIoService } from "./services/socketio.service";
+import { SocketIoGateway } from "./gateway/socketio.gateway";
+import { WebSocketGateway } from "./gateway/websocket.gateway";
+import { WebSocketService } from "./services/websocket.service";
+import { HttpModule } from "@nestjs/axios";
 
 @Module({
-  imports: [HttpModule, SocketIoModule, WebSocketModule],
+  imports: [HttpModule],
+  controllers: [HttpController],
+  providers: [
+    HttpService,
+    SocketIoGateway,
+    SocketIoService,
+    WebSocketGateway,
+    WebSocketService,
+  ],
+  exports: [SocketIoService, WebSocketService],
 })
 export class ProxyModule {}
