--- src/modules/proxy/socketio/socketio.module.ts
@@ -1,9 +1,0 @@
-import { Module } from "@nestjs/common";
-import { SocketIoGateway } from "./socketio.gateway";
-import { SocketIoService } from "./socketio.service";
-
-@Module({
-  providers: [SocketIoGateway, SocketIoService],
-  exports: [SocketIoService],
-})
-export class SocketIoModule {}
