--- src/modules/proxy/websocket/websocket.module.ts
@@ -1,9 +1,0 @@
-import { Module } from "@nestjs/common";
-import { WebSocketGateway } from "./websocket.gateway";
-import { WebSocketService } from "./websocket.service";
-
-@Module({
-  providers: [WebSocketGateway, WebSocketService],
-  exports: [WebSocketService],
-})
-export class WebSocketModule {}
